organized component folder
