+++ src/router/index.js
@@ -1,9 +1,9 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import CountriesGrid from '@/components/CountriesGrid'
-import CountryView from '@/components/CountryView'
-import SubdivisionsGrid from '@/components/SubdivisionsGrid'
-import SubdivisionView from '@/components/SubdivisionView'
+import CountriesGrid from '@/components/countries/CountriesGrid'
+import CountryView from '@/components/countries/CountryView'
+import SubdivisionsGrid from '@/components/subdivisions/SubdivisionsGrid'
+import SubdivisionView from '@/components/subdivisions/SubdivisionView'
 
 Vue.use(Router)
